Created list of valid mime types
If a certain download link's mime type is not valid, the download is skipped.

This is required because in origin download were advertised as 'application/x-download'
but now some of them are advertised as 'application/pdf' which is equally fine.
Not every mime type is to be accepted as sometimes empty files are advertised (and we
detect this by checking mime type)